Resolves a bug where watch would break output.
For whatever reason, this conditional would break the output of the module. I have no idea why this was added, but removing it does fix the issue. I've verified that this line is the culprit by removing it altogether, and in that case the build always fails, whether watch is on or not.

Resolves #79